Compiles in linux again, CmusikDir::OpenDir() should work as well.

+++ musikCore/src/musikPlayer.cpp
@@ -240,7 +240,7 @@ static void musikPlayerWorker( CmusikThread* thread )
 								recalc_steps = true;
 							}
 
-							if ( nMainStream != nChildCount )
+							if ( nMainStream != (int)nChildCount )
 							{
 								nMainStream = nChildCount;
 								recalc_steps = true;
